Fix MSVC compiler warnings for Windows build
This commit addresses three MSVC-specific warnings:

1. Variable shadowing in mode_controller.cpp (C4458)
   - Rename local variable 'properties' to 'objectProperties'
   - Prevents shadowing of Mode_Controller::properties class member

2. Variable shadowing in main.cpp (C4459)
   - Rename QTextStream 'cout' to 'qout'
   - Prevents shadowing of std::cout global declaration

3. Double to float32 truncation in ackermann_steer.cpp (C4305)
   - Add explicit static_cast<float> for SetTransform angle parameter
   - Box2D's b2Body::SetTransform expects float32 for angle

+++ veranda/Packages/builtin_plugin_logic/pkg-veranda_builtin_wheels/src/ackermann_steer.cpp
@@ -90,10 +90,10 @@ void Ackermann_Steer::_jointWheels()
 
         //Re-register the wheel bodies to get them positioned where they need to
         //be relative to the local origin
-        _rWheelBody->SetTransform(b2Vec2(_l1.get().toDouble()/2.0, 0), 90*DEG2RAD);
+        _rWheelBody->SetTransform(b2Vec2(_l1.get().toDouble()/2.0, 0), static_cast<float>(90*DEG2RAD));
         registerBody(_rWheelBody, {_rWheelModel});
 
-        _lWheelBody->SetTransform(b2Vec2(-_l1.get().toDouble()/2.0, 0), 90*DEG2RAD);
+        _lWheelBody->SetTransform(b2Vec2(-_l1.get().toDouble()/2.0, 0), static_cast<float>(90*DEG2RAD));
         registerBody(_lWheelBody, {_lWheelModel});
 
         //qDebug() << "Main body:" << _cBody->GetWorldCenter().x << _cBody->GetWorldCenter().y << _cBody->GetAngle();
